requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
twilio==8.10.0
supabase==1.0.4
httpx==0.25.1
pydantic==2.4.2
google-generativeai==0.3.1
python-multipart==0.0.6
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose==3.3.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
cryptography==41.0.5
uuid==1.30 